Refactor redux folder
