--- src/app/redux/reduxTypes.ts
@@ -1,10 +1,0 @@
-export interface Action<T> {
-	type: string;
-	payload: T;
-}
-
-export type Dispatch = (action: any) => any;
-
-export interface DispatchProps {
-	dispatch: Dispatch;
-}
